% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor.R
\name{ancestor}
\alias{ancestor}
\title{Calculate Ancestor Probability}
\usage{
ancestor(population = 10000, prev = 0.1)
}
\arguments{
\item{population}{The population size.}

\item{prev}{The previous probability.}
}
\value{
A vector of probabilities.
}
\description{
Calculates the probability of an ancestor given a population size and previous probability.
}
\examples{
ancestor(population = 10000, prev = 0.1)
}
