% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor.R
\name{ancestor.s}
\alias{ancestor.s}
\title{Simulate single ancestor}
\usage{
ancestor.s(population = 10000, prev = 0.1)
}
\arguments{
\item{population}{The population size.}

\item{prev}{The previous probability.}
}
\value{
A vector of probabilities.
}
\description{
Simulates an ancestor given a population size and prevalence.
}
\examples{
ancestor.s(population = 10000, prev = 0.1)
}
