% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryDAGs-package.R
\docType{package}
\name{ancestryDAGs-package}
\alias{ancestryDAGs}
\alias{ancestryDAGs-package}
\title{ancestryDAGs: DAGs ancestry quantification}
\description{
Package with simple functions to quantify DAGs in R. Based on course mmaterial from PhD course on Directed Acyclic Graphs (DAG) at AU.
}
\author{
\strong{Maintainer}: Andreas Gammelgaard Damsbo \email{agdamsbo@clin.au.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
