% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{child}
\alias{child}
\title{Calculate the probability of a child with given parameters}
\usage{
child(baserisk, por, parent)
}
\arguments{
\item{baserisk}{A numeric vector containing the base risk}

\item{por}{A list containing the odds ratios}

\item{parent}{A list containing the parent's risk factors}
}
\value{
A vector of probabilities of a child
}
\description{
Calculate the probability of a child with given parameters
}
\examples{
pop <- 10000
child(
  baserisk = 0.1,
  por = list(1.2, 0.8),
  parent = ancestor.multi(population = pop, prev = 0.2, n = 2)
)
}
